Add connection timeouts
Reduced the default long-polling duration from 120s to 50s, which appears to be the longest time Telegram will wait for.

Added connect and read timeouts, set to 5s and 50s respectively. This means the connection should timeout if it fails to contact Telegram within 5s, or connects but doesn't hear back from Telegram (regarding the long-polling) for 50s.

This should help recover the application if the connection to Telegram is broken (instead of waiting forever and needing a reboot to fix).